Miscellaneous fixes

+++ src/math/color3.cpp
@@ -17,7 +17,7 @@ export namespace xieite {
 		[[nodiscard]] friend bool operator==(const xieite::color3&, const xieite::color3&) = default;
 
 		[[nodiscard]] constexpr std::uint_least32_t value() const noexcept {
-			return (static_cast<std::uint_least32_t>(this->r) << 16) | (this->g << 8) | this->b;
+			return (static_cast<std::uint_least32_t>(this->r) << 16) | (static_cast<std::uint_least32_t>(this->g) << 8) | this->b;
 		}
 	};
 }
